Remove the code for selecting the HTTP2 asset server
The HTTP2 asset server has been disabled for a while now and will be removed in a future COGS version. This change removes the code which would select the HTTP2 server URL and means it will always use the standard HTTP1 asset server URLs.

There is a slight breaking change in this for TypeScript users as any calls to the `assetUrl` helper method with the boolean parameter will now produce a type error.